moved globals to global
