+++ config/global.js
@@ -27,6 +27,7 @@ const exp = {
 	NODE_ENV,
 	JWT_SHOW_TIMEOUT_WARNING_SECONDS,
 	JWT_TIMEOUT_SECONDS,
+	MAXIMUM_ALLOWABLE_TOTAL_ATTACHMENTS_SIZE_BYTE: (5 * 1024 * 1024), // 5MB
 };
 
 logger.info(util.inspect(exp, { depth: 1, compact: false }));
