--- config/globals.js
@@ -1,3 +1,0 @@
-module.exports = {
-	MAXIMUM_ALLOWABLE_TOTAL_ATTACHMENTS_SIZE_BYTE: (5 * 1024 * 1024), // 5MB
-};
